Move the MiroTransformable case class into the Miro transformer
