+++ catalogue_pipeline/transformer/transformer_miro/src/main/scala/uk/ac/wellcome/platform/transformer/miro/models/MiroTransformable.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.transformer.miro.models
+
+case class MiroTransformable(
+  sourceId: String,
+  MiroCollection: String,
+  data: String
+)
