+++ catalogue_pipeline/transformer/transformer_miro/src/main/scala/uk/ac/wellcome/platform/transformer/miro/modules/MiroTransformableModule.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.transformer.miro.modules
 import com.google.inject.{Provides, Singleton}
 import com.twitter.inject.{Injector, TwitterModule}
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.models.transformable.MiroTransformable
+import uk.ac.wellcome.platform.transformer.miro.models.MiroTransformable
 import uk.ac.wellcome.storage.ObjectStore
 import uk.ac.wellcome.storage.s3.S3StorageBackend
