+++ sbt_common/internal_model/src/main/scala/uk/ac/wellcome/models/transformable/Transformable.scala
@@ -8,12 +8,6 @@ import uk.ac.wellcome.models.transformable.sierra.{
   SierraItemRecord
 }
 
-case class MiroTransformable(
-  sourceId: String,
-  MiroCollection: String,
-  data: String
-)
-
 case class SierraTransformable(
   sierraId: SierraBibNumber,
   maybeBibRecord: Option[SierraBibRecord] = None,
